upgrade jedis、shiro-core.
upgrade jedis 4.0.0.
upgrade shiro-core 2.0.2.

+++ src/main/java/org/crazycake/shiro/common/WorkAloneRedisManager.java
@@ -3,8 +3,8 @@ package org.crazycake.shiro.common;
 import org.crazycake.shiro.IRedisManager;
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.JedisPoolConfig;
-import redis.clients.jedis.ScanParams;
-import redis.clients.jedis.ScanResult;
+import redis.clients.jedis.params.ScanParams;
+import redis.clients.jedis.resps.ScanResult;
 
 import java.util.HashSet;
 import java.util.List;
@@ -74,7 +74,7 @@ public abstract class WorkAloneRedisManager implements IRedisManager {
             jedis.set(key, value);
             // -1 and 0 is not a valid expire time in Jedis
             if (expireTime > 0) {
-                jedis.expire(key, expireTime);
+                jedis.expire(key, (long) expireTime);
             }
          } finally {
             jedis.close();
