+++ src/test/java/org/crazycake/shiro/RedisClusterManagerTest.java
@@ -38,10 +38,10 @@ public class RedisClusterManagerTest {
         verify(jedisCluster, times(0)).set(any(byte[].class), any(byte[].class));
         redisClusterManager.set(new byte[0], new byte[0], -1);
         verify(jedisCluster, times(1)).set(any(byte[].class), any(byte[].class));
-        verify(jedisCluster, times(0)).expire(any(byte[].class), any(int.class));
+        verify(jedisCluster, times(0)).expire(any(byte[].class), any(long.class));
         redisClusterManager.set(new byte[0], new byte[0], 0);
-        verify(jedisCluster, times(1)).expire(any(byte[].class), any(int.class));
+        verify(jedisCluster, times(1)).expire(any(byte[].class), any(long.class));
         redisClusterManager.set(new byte[0], new byte[0], 1);
-        verify(jedisCluster, times(2)).expire(any(byte[].class), any(int.class));
+        verify(jedisCluster, times(2)).expire(any(byte[].class), any(long.class));
     }
 }
